refactor: rewrite plugin in Typescript
