--- index.d.ts
@@ -1,11 +1,0 @@
-import { FastifyPluginAsync } from 'fastify';
-
-interface GraaspFileItemOptions {
-  /**
-   * Filesystem root path where the uploaded files will be saved
-   */
-  storageRootPath: string;
-}
-
-declare const plugin: FastifyPluginAsync<GraaspFileItemOptions>;
-export default plugin;
