+++ src/schema.ts
@@ -1,0 +1,19 @@
+
+const upload = {
+  querystring: {
+    type: 'object',
+    properties: {
+      parentId: { $ref: 'http://graasp.org/#/definitions/uuid' }
+    },
+    additionalProperties: false
+  }
+};
+
+const download = {
+  params: { $ref: 'http://graasp.org/#/definitions/idParam' },
+};
+
+export{
+  upload,
+  download
+}
